Create utils file for FIrebase utiliity funcs
